requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
backports-datetime-fromisoformat==2.0.1
beautifulsoup4==4.12.3
bilix==0.18.8
blinker==1.8.2
browser-cookie3==0.19.1
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
construct==2.8.8
contourpy==1.3.0
cycler==0.12.1
danmakuC==0.3.6
decorator==5.1.1
exceptiongroup==1.2.2
Flask==3.0.3
fonttools==4.53.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
imageio==2.35.1
imgaug==0.4.0
itsdangerous==2.2.0
jeepney==0.8.0
Jinja2==3.1.4
json5==0.9.25
kiwisolver==1.4.5
lazy_loader==0.4
lz4==4.3.3
m3u8==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
networkx==3.3
numpy==1.23.5
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==24.1
paddle-bfloat==0.1.7
paddlepaddle==2.6.1
pillow==10.4.0
protobuf==5.28.0
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymp4==1.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tifffile==2024.8.28
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
